refactor(measurements): reorganize measurement package
