+++ diode_measurement/core/measurement.py
